fix: resolve circular dependency in tool context
Move tool execution context to separate toolContext.ts file to avoid
circular dependency between manager.ts and Read.ts that caused
TOOL_DEFINITIONS to be undefined at startup.
